# openapi.yaml
openapi: 3.0.3
info:
  title: Payment Dashboard API
  version: "1.0.0"
  description: |
    Internal dashboard API for monitoring and reviewing incoming payments.

    Two roles:
    - `cs`: read-only access to payments and summary.
    - `operation`: same as `cs` plus the ability to approve/reject payments
      that are still in `processing` state.

    All `/dashboard/*` endpoints (except login) require a Bearer JWT issued
    by `POST /dashboard/v1/auth/login`.
servers:
  - url: http://localhost:8080
    description: Local backend (make run / docker compose)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP-aligned numeric error code.
          example: 401
        message:
          type: string
          example: "unauthorized: missing or invalid token"
        details:
          type: object
          additionalProperties: true
          description: Optional structured details (validation errors, etc.).
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "operation@test.com"
        password:
          type: string
          format: password
          minLength: 1
          example: "password"
    User:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [cs, operation]
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: "JWT bearer token. Send as `Authorization: Bearer TOKEN`."
        user:
          $ref: '#/components/schemas/User'
...
  responses:
    BadRequest:
      description: Request payload or query parameters failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
  /dashboard/v1/auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      security: []
      summary: Login with email + password
      description: Returns a JWT and the authenticated user. Public endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
...